Copy JS files to public dir when building assets
By default this copies to the /frontend/dist directory which meant the
uglified JS was not being served.

By default lineman runs copy:dev as part of it's common tasks
(confusingly really given it's *:dev name) which copies the non
minified version over which is how this problem didn't have more drastic
effects.