Add source maps to dist JS
This adds the generated source map from the concat process as an source
map input for uglify, which generates a source map output.

We need to override the URL that uglify outputs on the source map as by
default it outputs to current directory which would be
`/dist/js/app.js.map`.

I have a slight concern that ngmin could mess up the source maps
somewhat, since that runs between concat and uglify to change the
generated JS. It didn't appear to from browsing the output though.