Configure track documentation image path
The track documentation often needs to contain images, and those
images live in the track repository. This means that when the
documentation refers to "/docs/img/test.png", that is in the
track repository, not in the Exercism repository.

When we parse the markdown, the HTML creates an image with a source
that has a URL which is relative to the Exercism site.

We need to rewrite the image path to use the API endpoint which serves
track images.