Add dictionary with archive tooltip as first entry
There is some confusion around what the archive functionality actually
does, and also some concern about maintaining consistency across
tooltip text.

This commit proposes making use of the i18n gem as it ships with
activesupport and will be intuitive to those from a rails background.
It would also support providing translations at some point in future if
that were ever required.

It solves issue:
https://github.com/exercism/exercism.io/issues/3349

This is my first (ever) PR to an open source project, encouraged by
Katrina's appearance on the Bike Shed podcast, but you won't hurt my
feelings if there are things that are not correct.