Remove tooltips in favour of copy on dashboard
This commit removes the tooltips completely in favour of explaining what
archiving does on the dashboard.

I prefer this approach as while I expect most of Exercism's traffic to
be desktop (keyboard and mouse) tooltips have very poor usability on
touch screens such as tablets and mobiles.

It solves issue: