Display a flash notice from ClientVersion
This is defined in ExercismWeb::Routes::Profile, because it's the most
logical subclass of ExercismWeb::Routes::Core that runs on every normal
web page request. Defining this before filter in Core itself causes the
before filter to run once for each of its many subclasses that load for
the page: generally each router that has a partial displayed. Sinatra!
Since the user profile nav appears on every page, this is a candidate
that works.

Also, requests for /icon come through the normal rack stack, and each of
those requests also runs before filters. I want this to run and set the
flash notice once per page, so this also checks the HTTP_ACCEPT header
to see whether or not a text/html page is being loaded.