Alphabetize exercism.rb requires
This helps make it clear where to put new requires, and distributes new
requires in such a way that future conflicts are less likely.

All the tests pass still...

Any load-time dependency order requirements should theoretically blow up
immediately during the require. That wouldn't hold true if some class
were using metaprogramming to search for existing constants to create a
list of things matching a pattern or something, but are we doing that?

@kytrinyx, is there any reason you can think of that reordering here may
cause problems?