Use a class instead of id for write_tab
This is thinking ahead to a future change when we may have multiple
copies of these tabs on the page at once. Since this branch is about to
start referencing the write tab, making this change now will make the
future change easier by creating fewer things that need to change.