Adding instructions for Windows Subsystem for Linux
This pull request adds steps for Windows Subsystems for Linux (Bash on
Ubuntu on Windows), and modifies one linux step for getting Node.js
package manger NPM.

This may be updated in the future as I find out more things, but these
are the steps I used to get bin/setup to run correctly.

This also modifies changes I would have made in my previous pull request
to add a step if you are having trouble getting the current required
version of ruby (2.3.3).