Push footer to bottom of the screen.
fixes #3465

* Why was this change necessary?
The footer is static and hence if the page that the user is viewing
doesn't take up the entire viewport, the footer appears somewhere in the
middle of the page. This is not the expected behaviour, ideally we want
the footer to appear at the very end of the viewport.

* How does it address the problem?
The `.navbar-fixed-bottom` class from Bootstrap was utilized to fix the
footer to the bottom of the page. Additionally, to make sure that the
footer didn't overlap any content on the page, a `padding-bottom: 300px`
was applied to the `body`.

* Are there any side effects?
I noticed that we have a media query that increases the `height` of the
footer by `20px` somewhere in the 1000 - 1200px range (tablet range).
This makes the additional `padding-bottom` that was applied to the body,
appear inconsistent across all screen sizes. For example, the spacing
between the footer and body in tablet mode is more than desktop
mode.

Unfortunately I don't have the required css skills to come up with a
more robust and scalable solution so please feel free to suggest
alternatives!