Include missing data points for 0 values
Because of the nature of the counting query, users who submitted 0
comments during a period were not represented in the statistics.
This was apparent because the number of comments per user in each period
was never reported to be less than 1.

This is difficult to accomplish in the original counting query, so
instead this patch fills 0s for each participating user (who commented
at least once during the experiment period) who isn't already
represented by a comment count.