Handle dirty membership data
I'm still not sure how this happens, but we sometimes get a membership invite (confirmed or not),
that still has a membership invitation lying around.

This was raising an exception and failing to confirm the invite.

This adds tests for both of the cases, and does the right thing, adding the user
to the team and deleting the invite.