Add db:migrate step to db:reset
When resetting the database you always want to run the db:migrate task
to update to the latest schema.
This patch adds the migrate step to reset so this will happen
automatically for you.

It will then helpfully suggest the next step, which is seeding the database.
Unless you are resetting the testing database, where seeding is undesirable.

If you wanted to avoid the migration step you can still run:
```
rake db:drop db:create
```