Add test coverage to track documentation
This has a fair amount of complexity, and I want to refactor it. The test coverage
was not good enough.

I removed the tests at the track presenter level, since that will likely
be going away, and because we're testing very thoroughly at the route
level.

I added some skipped tests for things that I want to fix, but I didn't
want to mix it into the same pull request, as this change should be
purely about adding tests, not changing behavior.