Rerun lineman when coffee/js assets change
This is a hack. Lineman really ought to instead be doing its own
reloading. It watches and compiles the files but doesn't put them in the
right places to be effective. I have not spent the time to learn lineman
well enough to fix it.

Instead as a workaround, restart lineman, because starting lineman puts
everything in its right place.

See https://github.com/exercism/exercism.io/issues/2936#issuecomment-315500149