Prevent reflected XSS via unescaped HTML tags in comment history
Since exercism doesn't implement CSRF protection, a malicious site could
submit comments on behalf of a user containing <script> tags that would
execute when rendering that user's comment history. That script could
send the user's cookie to a malicious party, who could then impersonate
the user on exercism.

This patch escapes HTML in comment history, which in addition to being a
security risk can interfere with the proper function of the page.