Fix Navigation Bar on iPad (Tablet)
It was reported in https://github.com/exercism/exercism.io/issues/3785
that the nav bar was not rendering properly on the iPad.

Looking at the scss rules it looked like the rule to collapse the nav
bar was set to 768px which is the breakpoint for tablets in media
queries. Since it was within that range, that resolution was not
considered when determining if it should collapse the nav bar or not.

Bumping it up to the next breakpoint now collapses the items when
rendered from a tablet.

This resolves https://github.com/exercism/exercism.io/issues/3785