Notify user when no exercises are selected when archiving
Resolves https://github.com/exercism/exercism.io/issues/3786

A discussion was started in https://github.com/exercism/exercism.io/issues/3786 on how to address this issue. It was decided that we should show the user a flash message when no exercises are selected when trying to click on the archive button.

The approach taken here is to catch the `ActiveRecord::RecordNotFound` exception and display a message to the user. The approach that was discarded was adding validation in javascript to prevent the user from clicking on the button to *Archive* if no exercises were selected.