Add description to 'stale' label
We decided to use the status/stale label to indicate that we think
something is no longer a problem. We will comment in the issue thread
when marking something as stale, asking people to double-check.

If we don't hear anything back, we will close the issue.
This is not an automated process at the moment.